chore: Update package version to 1.5.3 and enhance Chainflip API endpoints
This commit updates the project version in package.json to 1.5.3. It also expands the chainIdToAlchemyNetwork mapping in covalentApi.ts to include additional networks. Furthermore, the API endpoints for Chainflip in SKClientProviderManager.js and client.js have been updated to use HTTPS for improved reliability. These changes aim to enhance the overall robustness and maintainability of the codebase.